+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 - Mocked Cache test on controller
 ### Fixed
 - Fix deleting cache when template is deleted
+
+## [1.0.2] 2018-11-13
+
+### Added
+- New field scope in templates entity
+- Test support for field scope in backend and controllers